requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
browserstack-local==1.2.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
h11==0.12.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
outcome==1.1.0
percy-selenium==1.0.4
psutil==5.8.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.18.0
requests==2.25.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.6
Werkzeug==2.0.1
wsproto==1.0.0
